update variables for loadPEPdb and LOLArun test
I'm changing the names of the variables for the PEP containing only BED files. This is done to test the PEP loading function and make sure I can run the `LOLA` analysis. These changes are not definitive as I need to find a consensus for the annotation sheet required by the `bed_maker` pipeline.